fix: Better org and docs
Moves the multiprefix logic to the `core/` and updates
some documentation.
