+++ src/core/mod.rs
@@ -2,5 +2,6 @@
 pub mod composer;
 pub mod convert_handler;
 pub mod friendly_tekton;
+pub mod multiprefix_tekton;
 pub mod snipmate_tekton;
 pub mod sort_handler;
