+++ src/models/mod.rs
@@ -1,5 +1,5 @@
 //! The structs to model FriendlySnippets or Snipmate snippets
 pub mod args;
 pub mod friendly;
-pub mod multi_prefix_friendly;
+pub mod multiprefix_friendly;
 pub mod snipmate;
